rename geography header
